Remove p_test variable dependencies on package p (#857)
* Remove h-fam reference

* Remove p_test variable dependencies on package p
