+++ ytypes/list_test.go
@@ -524,6 +524,87 @@ func TestUnmarshalUnkeyedList(t *testing.T) {
 	}
 }
 
+func TestUnmarshalSingleListElement(t *testing.T) {
+	listSchema := &yang.Entry{
+		Name:     "struct-list",
+		Kind:     yang.DirectoryEntry,
+		ListAttr: yang.NewDefaultListAttr(),
+		Dir: map[string]*yang.Entry{
+			"leaf-field": {
+				Kind: yang.LeafEntry,
+				Name: "leaf-field",
+				Type: &yang.YangType{Kind: yang.Yint32},
+			},
+			"enum-leaf-field": {
+				Kind: yang.LeafEntry,
+				Name: "enum-leaf-field",
+				Type: &yang.YangType{Kind: yang.Yenum},
+			},
+			"leaf2-field": {
+				Kind: yang.LeafEntry,
+				Name: "leaf2-field",
+				Type: &yang.YangType{Kind: yang.Yint64},
+			},
+		},
+	}
+
+	type ListElemStruct struct {
+		LeafName  *int32   `path:"leaf-field"`
+		EnumLeaf  EnumType `path:"enum-leaf-field"`
+		Leaf2Name *int64   `path:"leaf2-field"`
+	}
+
+	tests := []struct {
+		desc    string
+		schema  *yang.Entry
+		json    string
+		parent  any
+		want    any
+		wantErr string
+	}{
+		{
+			desc:   "success",
+			schema: listSchema,
+			json:   `{ "leaf-field" : 42, "enum-leaf-field" : "E_VALUE_FORTY_TWO"}`,
+			parent: &ListElemStruct{
+				Leaf2Name: ygot.Int64(42),
+			},
+			want: &ListElemStruct{
+				LeafName:  ygot.Int32(42),
+				Leaf2Name: ygot.Int64(42),
+				EnumLeaf:  42,
+			},
+		},
+		{
+			desc:    "bad field",
+			schema:  listSchema,
+			json:    `{ "leaf-field" : 42, "bad-field" : "E_VALUE_FORTY_TWO"}`,
+			parent:  &ListElemStruct{},
+			wantErr: `parent container struct-list (type *ytypes.ListElemStruct): JSON contains unexpected field bad-field`,
+		},
+	}
+
+	var jsonTree interface{}
+	for _, tt := range tests {
+		t.Run(tt.desc, func(t *testing.T) {
+			if err := json.Unmarshal([]byte(tt.json), &jsonTree); err != nil {
+				t.Fatalf("%s : %s", tt.desc, err)
+			}
+
+			err := Unmarshal(tt.schema, tt.parent, jsonTree)
+			if diff := errdiff.Text(err, tt.wantErr); diff != "" {
+				t.Fatalf("%s: Unmarshal error not expected:\n%s", tt.desc, diff)
+			}
+			if err == nil {
+				got, want := tt.parent, tt.want
+				if diff := cmp.Diff(want, got); diff != "" {
+					t.Errorf("%s: Unmarshal (-want, +got):\n%s", tt.desc, diff)
+				}
+			}
+		})
+	}
+}
+
 func TestUnmarshalStructKeyedList(t *testing.T) {
 	containerWithLeafListSchema := &yang.Entry{
 		Name: "container",
